Streamlined the extract_*_array functions to eliminate the C++ code.
The C++ code was only required to perform expansion of duplicated
indices in extract_array(), but this is more easily handled by natively
extracting an array from TileDB and then expanding it in R. For sparse
arrays, we just convert the immediate output from TileDB into a dense
array first, and then expand it, avoiding the need for extra code.

For extract_sparse_array(), we don't even need to consider the
possibility of duplicates, which allows for further simplification.