Small code simplification
nzvals/nzwhich work on any SparseArray derivative (i.e. COO_SparseArray
and SVT_SparseArray objects). They even work on dgCMatrix objects and
other sparse representations from the Matrix package.

Furthermore, nzdata() and nzcoo() are just accessors for the 'nzdata'
and 'nzcoo' slots of a COO_SparseArray object, but these slots are not
guaranteed to contain the nonzero values and their index in an order
that is consistent with the "first dimension is the fastest moving
dimension" convention. OTOH nzvals/nzwhich are guaranteed to return the
nonzero values and their index in an order that follows this convention.
Just a FYI as I don't know if this is relevant in the context of the
write_block() method for TileDBRealizationSink objects.